Adds ability to inject custom metadata at client registration
Fixes gh-1172
